Refactoring and feature additions
- Added a color temp button type for Quasar Science Rainbow 1 lights.
  Their native color temp range is smaller than the Aeon 2 lights, so
  the settings need to be tweaked.
- Refactored a bunch of things to make the Aeon lights their own class,
  so you can create a light rather than a bunch of independent
  controls.

Next up, the Rainbow lights will need to get the same treatment as the
Aeons, but I'll probably want to refactor a bunch of this again, as 95%
of the logic will be shared between them, with the only differences
(for now, at least), being the color temp settings.

Going forward, the Rainbows (and a Rainbow 2, which is on order, and will
need its own class as well) will probably want to support a handful of
additional settings, as they support multiple color modes and a set of
special "FX" settings which might be useful once in a blue moon.

Thinking about this, I should probably break the light UI and the light
backend apart, and make it so that it's possible to have multiple lights
(even different types?) wired into a single Loupedeck control knob.
That'll be entertaining when color temps don't line up, but otherwise
is probably reasonable.  Each light will define a min/max color temp
setting, and I can change the meaning of a.ColorTemp to be Kelvin, rather
than the raw DMX value.  The conversion of Kelvin->DMX will be a function
of the light.

Then I can just do something like this:

  light1 := NewAeon2Light(udmx, 1)
  light2 := NewAeon2Light(udmx, 3)
  lightcontrol1 := NewLightControl(l, LeftTop, [light1, light2])

for 1:1 control:light cases, that shortens to

  lightcontrol2 := NewLightControl(l, LeftMiddle, NewAeon2Light(udmx, 5))

Signed-off-by: Scott Laird <scott@sigkill.org>