Added LightController and per-light classes.
This split leaves a generic light controller that handles brightness and color temp, plus a per-light class that
handles brightness and color setting for that type of light.

This opens the door for non-DMX lights, and also for multiple lights attached to one controller.

In addition, this abstracts color temperature; before color temp was really just an uint,
and each type of light had a struct that listed possible (dmx value, color name, RGB for loupedeck)
settings.  Now, the controller handles color temps as degrees kelvin, and each individual light class
can map that to whatever it's able to handle.  Each light class exports a min and max color temp,
and the controller should handle mixed light sets just fine.  There's still a bit more work needed to get
a good UI for really wide-range lights (like the Quasar Rainbow 2, which can do ~1500k-10000k), but it'll
work now.

It should also be trivial to create single-temp lights.  The UI will still show a color temp button for them, but it'll have a single value and you won't be able to change it.  In addition, if you group a single-color light with a variable-color light in a single LightController, they should both get the same temp.

Next up:
- Split into multiple files
- Add tests.  The color code should be easy enough to test, once it's not in `package main`.
- Add support for Quasar Science Rainbow (1) lights in mode 12 or 13 (need to see which one I'm using, don't remember off the top of my head) with 16-bit brightness and special effects on different DMX IDs.
- Add support for Quasar Science Rainbow 2 lights, with a wider color range and hopefully a wider brightness range.
- Add support for saving and loading color settings to a config file.
- Add support for driving a Blackmagic Hyperdeck via the Loupedeck CT's dial and screen.  I want to be able to dial through background videos and start/pause them that way.

Signed-off-by: Scott Laird <scott@sigkill.org>